rest/requirements.txt
fastapi[standard]
pydantic
psycopg[binary]
pyjwt<2.10
passlib[bcrypt]